revert: restore hand-written SDKs, remove codegen workflow
- Restored 11 hand-written SDKs from commit 7592591a
- Removed openapi-codegen.yml (would overwrite custom SDK code)
- Kept pgcrypto fix and custom_headers migration in DB

+++ sdks/swift/OpenAPIClient/Classes/OpenAPIs/Models/SubscriptionResponse.swift
@@ -18,9 +18,9 @@ public struct SubscriptionResponse: Codable, JSONEncodable, Hashable {
     public var webhookLimit: Int
     public var endpointLimit: Int
     public var retentionDays: Int
-    public var monthlyPriceCents: Int64
+    public var monthlyPriceCents: Int
 
-    public init(plan: String, status: String, paymentProvider: String, webhookLimit: Int, endpointLimit: Int, retentionDays: Int, monthlyPriceCents: Int64) {
+    public init(plan: String, status: String, paymentProvider: String, webhookLimit: Int, endpointLimit: Int, retentionDays: Int, monthlyPriceCents: Int) {
         self.plan = plan
         self.status = status
         self.paymentProvider = paymentProvider
